codegen: adjust types of ImS16

+++ cmd/codegen/arguments_wrapper.go
@@ -69,7 +69,7 @@ func getArgWrapper(
 		"ImS8":                simpleW("int", "C.ImS8"),
 		"ImS8*":               simplePtrW("C.ImS8", "int8"),
 		"const ImS8*":         simplePtrSliceW("C.ImS8", "int8"),
-		"ImS16":               simpleW("int", "C.ImS16"),
+		"ImS16":               simpleW("int16", "C.ImS16"),
 		"ImS16*":              simplePtrW("int16", "C.ImS16"),
 		"const ImS16*":        simplePtrSliceW("C.ImS16", "int"),
 		"ImS32":               simpleW("int", "C.ImS32"),
